+++ test/SubstanceProviderTest.php
@@ -10,6 +10,8 @@ use PHPUnit\Framework\Attributes\CoversMethod;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseFactoryInterface;
+use SubstancePHP\Container\Container;
+use SubstancePHP\HTTP\Application;
 use SubstancePHP\HTTP\ContextFactoryInterface;
 use SubstancePHP\HTTP\Environment;
 use SubstancePHP\HTTP\EnvironmentInterface;
@@ -18,6 +20,7 @@ use SubstancePHP\HTTP\Middleware\BodyParserMiddleware;
 use SubstancePHP\HTTP\Middleware\RouteActorMiddleware;
 use SubstancePHP\HTTP\Middleware\RouteMatcherMiddleware;
 use SubstancePHP\HTTP\SubstanceProvider;
+use TestUtil\Fixture\ContentTypeOverrideProvider;
 
 #[CoversClass(SubstanceProvider::class)]
 #[CoversMethod(SubstanceProvider::class, 'factories')]
@@ -42,5 +45,43 @@ class SubstanceProviderTest extends TestCase
         $this->assertArrayHasKey(ResponseFactoryInterface::class, $result);
         $this->assertArrayHasKey(RouteActorMiddleware::class, $result);
         $this->assertArrayHasKey(RouteMatcherMiddleware::class, $result);
+        $this->assertArrayHasKey('substance.http.default-content-type', $result);
+    }
+
+    #[Test]
+    public function defaultContentType(): void
+    {
+        $environment = new Environment([]);
+        $container = Container::from(SubstanceProvider::factories($environment));
+        $actual = $container->get('substance.http.default-content-type');
+        $expected = 'text/html; charset=utf-8';
+        $this->assertSame($expected, $actual);
+    }
+
+    #[Test]
+    public function defaultContentTypeOverridable(): void
+    {
+        $env = [];
+        $actionRoot = \implode(
+            \DIRECTORY_SEPARATOR,
+            [\dirname(__DIR__), 'testutil', 'fixture', 'action'],
+        );
+        $templateRoot = \implode(
+            \DIRECTORY_SEPARATOR,
+            [\dirname(__DIR__), 'testutil', 'fixture', 'template'],
+        );
+        $instance = Application::make(
+            env: $env,
+            actionRoot: $actionRoot,
+            templateRoot: $templateRoot,
+            providers: [
+                SubstanceProvider::class,
+                ContentTypeOverrideProvider::class,
+            ],
+            middlewares: [],
+            htmlEncoding: 'utf-8',
+        );
+
+        $this->assertSame('application/json', $instance->get('substance.http.default-content-type'));
     }
 }